Refactor validators to smaller files
