+++ src/json-schema/utils/index.js
@@ -1,7 +1,8 @@
 import angular from 'angular';
 
 import { getValidModelParts } from './valid-model';
-import { getValidationFailures, isValidSchema } from './validation';
+import { getValidationFailures } from './validation-failures';
+import { isValidSchema } from './schema-validators';
 
 function Utils() {
   this.getValidModelParts = getValidModelParts;
